# app_books/src/main/java/com/programacion/distribuida/books/dto/BookDto.java
package com.programacion.distribuida.books.dto;

import lombok.Data;

import java.math.BigDecimal;

@Data
public class BookDto {

    private Integer id;

    private String isbn;

    private String title;

    private BigDecimal price;

    private String author;
}